+++ arduino/defs.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#define RFID_VCC_PIN 3
 #define LED_CONFIRM_PIN  4
 #define BUZZER_PIN 5
 #define RFID_SS_PIN 10